+++ internal/service/credentials/resource_credential_issuer_profile.go
@@ -70,8 +70,8 @@ func (r *CredentialIssuerProfileResource) Schema(ctx context.Context, req resour
 		Attributes: map[string]schema.Attribute{
 			"id": framework.Attr_ID(),
 
-			"environment_id": framework.Attr_LinkID(framework.SchemaDescription{
-				Description: "TThe ID of the environment to create the credential issuer in."},
+			"environment_id": framework.Attr_LinkID(
+				framework.SchemaAttributeDescriptionFromMarkdown("TThe ID of the environment to create the credential issuer in."),
 			),
 
 			"application_instance_id": schema.StringAttribute{