+++ internal/service/sso/data_source_population.go
@@ -66,8 +66,8 @@ func (r *PopulationDataSource) Schema(ctx context.Context, req datasource.Schema
 		Attributes: map[string]schema.Attribute{
 			"id": framework.Attr_ID(),
 
-			"environment_id": framework.Attr_LinkID(framework.SchemaDescription{
-				Description: "The ID of the environment that is configured with the population."},
+			"environment_id": framework.Attr_LinkID(
+				framework.SchemaAttributeDescriptionFromMarkdown("The ID of the environment that is configured with the population."),
 			),
 
 			"population_id": schema.StringAttribute{